+++ CHANGELOG.md
@@ -34,7 +34,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Replaced babel-instrumenter-loader with babel-plugin-istanbul.
 - @aofl/router - match-route-middleware caches the response after matching the route instead of the route.
 - @aofl/web-components/aofl-element export an object containing AoflElement and all lit-element exports
+- @aofl/web-components/aofl-element uses css`` to render styles
 - Updated all references to AoflElement to reflect updated module export
+- babel settings
 
 ### [Deprecated]
 ### [Removed]